Corrected homework
