+++ Bulldozer.java
@@ -1,0 +1,11 @@
+package core.basesyntax;
+public class Bulldozer extends Machine {
+    @java.lang.Override
+    public void doWork() {
+        System.out.println("Bulldozer started work" );
+    }
+    @java.lang.Override
+    public void stopWork() {
+        System.out.println("Bulldozer stopped work");
+    }
+}
